♻️ refactor(webui): plugin-config 状态机下沉为领域 hook
候选 3：plugin-config.tsx 2108→1342 行。config/sourceCode 是可编辑草稿，按 model 范式只搬状态机、不转 Query。

- 新增 plugin-config/hooks/usePluginList（列表+搜索/过滤+选中+深链接 ?plugin=+市场版本比对+启停+派生）
- usePluginConfigEditor（visual/source 双模式草稿+schema+保存+重置+TOML 校验+useBlocker 离开拦截；hasChanges 改为纯派生消除 set-state-in-effect）
- usePluginLifecycle（更新/卸载+pluginProgressClient WS 进度订阅，useEffect 订阅/cleanup 退订）
- utils 抽出共用 getNestedRecord/setNestedField
- 子组件 CodeEditor/ListFieldEditor/SectionRenderer/FieldRenderer 零改动
- 特征化测试 5/5 全绿
