+++ dashboard/src/routes/plugin-config.tsx
@@ -1,5 +1,4 @@
-import { useBlocker } from '@tanstack/react-router'
-import { useState, useEffect, useCallback } from 'react'
+import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { ScrollArea } from '@/components/ui/scroll-area'
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
@@ -18,7 +17,6 @@ import { ListFieldEditor } from '@/components/ListFieldEditor'
 import { Alert, AlertDescription } from '@/components/ui/alert'
 import { CodeEditor } from '@/components/CodeEditor'
 import { useTheme } from '@/components/use-theme'
-import { parse as parseToml } from 'smol-toml'
 import {
   Select,
   SelectContent,
@@ -57,32 +55,20 @@ import {
   Layout,
   Trash2,
 } from 'lucide-react'
-import { useToast } from '@/hooks/use-toast'
 import { RestartProvider, useRestart } from '@/lib/restart-context'
 import { RestartOverlay } from '@/components/restart-overlay'
-import {
-  fetchPluginList,
-  getInstalledPlugins,
-  getPluginConfigSchema,
-  getPluginConfig,
-  getPluginConfigRaw,
-  updatePluginConfig,
-  updatePluginConfigRaw,
-  resetPluginConfig,
-  togglePlugin,
-  uninstallPlugin,
-  updatePlugin,
-  type InstalledPlugin,
-  type PluginLoadProgress,
-  type PluginConfigSchema,
-  type ConfigFieldSchema,
-  type ConfigSectionSchema,
-  type ItemFieldDefinition,
+import type {
+  InstalledPlugin,
+  ConfigFieldSchema,
+  ConfigSectionSchema,
+  ItemFieldDefinition,
 } from '@/lib/plugin-api'
-import type { PluginInfo } from '@/types/plugin'
-import { pluginProgressClient } from '@/lib/plugin-progress-client'
 import { PluginIcon } from './plugins/PluginIcon'
 import { getPluginTypeLabel } from './plugins/types'
+import { getNestedRecord } from './plugin-config/utils'
+import { usePluginList } from './plugin-config/hooks/usePluginList'
+import { usePluginLifecycle } from './plugin-config/hooks/usePluginLifecycle'
+import { usePluginConfigEditor } from './plugin-config/hooks/usePluginConfigEditor'
 
 // 字段渲染组件
 interface FieldRendererProps {
@@ -90,16 +76,6 @@ interface FieldRendererProps {
   value: unknown
   onChange: (value: unknown) => void
   sectionName: string
-}
-
-type PluginStatusIcon = 'loading' | 'warning' | 'circuit'
-
-interface PluginStatusMeta {
-  dotClassName: string
-  label: string
-  badgeClassName?: string
-  icon?: PluginStatusIcon
-  showsBadge?: boolean
 }
 
 function getLocaleCandidates(language: string): string[] {
@@ -167,56 +143,6 @@ function localizeItemFields(
       },
     ])
   )
-}
-
-function getNestedRecord(config: Record<string, unknown>, path?: string): Record<string, unknown> | undefined {
-  if (!path) {
-    return undefined
-  }
-  const parts = path.split('.').filter(Boolean)
-  let current: unknown = config
-
-  for (const part of parts) {
-    if (!current || typeof current !== 'object' || Array.isArray(current)) {
-      return undefined
-    }
-    current = (current as Record<string, unknown>)[part]
-  }
-
-  if (!current || typeof current !== 'object' || Array.isArray(current)) {
-    return undefined
-  }
-
-  return current as Record<string, unknown>
-}
-
-function setNestedField(
-  config: Record<string, unknown>,
-  path: string,
-  fieldName: string,
-  value: unknown,
-): Record<string, unknown> {
-  const parts = path.split('.').filter(Boolean)
-  const nextConfig: Record<string, unknown> = { ...config }
-  let currentTarget = nextConfig
-  let currentSource: Record<string, unknown> | undefined = config
-
-  for (const part of parts) {
-    const sourceValue: unknown = currentSource?.[part]
-    const nextValue =
-      sourceValue && typeof sourceValue === 'object' && !Array.isArray(sourceValue)
-        ? { ...(sourceValue as Record<string, unknown>) }
-        : {}
-    currentTarget[part] = nextValue
-    currentTarget = nextValue
-    currentSource =
-      sourceValue && typeof sourceValue === 'object' && !Array.isArray(sourceValue)
-        ? (sourceValue as Record<string, unknown>)
-        : undefined
-  }
-
-  currentTarget[fieldName] = value
-  return nextConfig
 }
 
 function FieldRenderer({ field, value, onChange }: FieldRendererProps) {
@@ -423,7 +349,7 @@ function SectionRenderer({ sectionName, section, config, onChange }: SectionRend
   const sectionConfig = getNestedRecord(config, resolvedSectionName)
   const title = resolveLocalizedText(section.title, language, sectionName, section.i18n, 'title')
   const description = resolveLocalizedText(section.description, language, '', section.i18n, 'description')
-  
+
   // 按 order 排序字段
   const sortedFields = Object.entries(section.fields)
     .filter(([, field]) => !field.hidden)
@@ -480,237 +406,38 @@ interface PluginConfigEditorProps {
 }
 
 function PluginConfigEditor({ plugin, onBack, initialTab }: PluginConfigEditorProps) {
-  const { toast } = useToast()
   const { triggerRestart, isRestarting } = useRestart()
   const { i18n } = useTranslation()
   const language = i18n.resolvedLanguage || i18n.language || 'zh'
-  const [editMode, setEditMode] = useState<'visual' | 'source'>('visual')
-  const [pluginPageTab, setPluginPageTab] = useState<'settings' | 'details'>('settings')
-  const [schema, setSchema] = useState<PluginConfigSchema | null>(null)
-  const [activeConfigTab, setActiveConfigTab] = useState<string | undefined>(initialTab)
-  const [config, setConfig] = useState<Record<string, unknown>>({})
-  const [originalConfig, setOriginalConfig] = useState<Record<string, unknown>>({})
-  const [sourceCode, setSourceCode] = useState('')
-  const [originalSourceCode, setOriginalSourceCode] = useState('')
-  const [loading, setLoading] = useState(true)
-  const [saving, setSaving] = useState(false)
-  const [hasChanges, setHasChanges] = useState(false)
-  const [hasTomlError, setHasTomlError] = useState(false)
-  const [resetDialogOpen, setResetDialogOpen] = useState(false)
-  const [internalLeavePromptOpen, setInternalLeavePromptOpen] = useState(false)
 
-  const navigationBlocker = useBlocker({
-    shouldBlockFn: () => hasChanges,
-    enableBeforeUnload: hasChanges,
-    withResolver: true,
-  })
-
-  // 加载配置
-  const loadConfig = useCallback(async () => {
-    setLoading(true)
-    try {
-      const [schemaResult, configResult, rawResult] = await Promise.all([
-        getPluginConfigSchema(plugin.id),
-        getPluginConfig(plugin.id),
-        getPluginConfigRaw(plugin.id)
-      ])
-      
-      if (!schemaResult.success) {
-        toast({
-          title: '加载配置架构失败',
-          description: schemaResult.error,
-          variant: 'destructive'
-        })
-        return
-      }
-      
-      if (!configResult.success) {
-        toast({
-          title: '加载配置数据失败',
-          description: configResult.error,
-          variant: 'destructive'
-        })
-        return
-      }
-      
-      if (!rawResult.success) {
-        toast({
-          title: '加载原始配置失败',
-          description: rawResult.error,
-          variant: 'destructive'
-        })
-        return
-      }
-      
-      setSchema(schemaResult.data)
-      setConfig(configResult.data)
-      setOriginalConfig(JSON.parse(JSON.stringify(configResult.data)))
-      setSourceCode(rawResult.data)
-      setOriginalSourceCode(rawResult.data)
-    } catch (error) {
-      toast({
-        title: '加载配置失败',
-        description: error instanceof Error ? error.message : '未知错误',
-        variant: 'destructive'
-      })
-    } finally {
-      setLoading(false)
-    }
-  }, [plugin.id, toast])
-
-  useEffect(() => {
-    loadConfig()
-  }, [loadConfig])
-
-  // 检测配置变化
-  useEffect(() => {
-    if (editMode === 'visual') {
-      setHasChanges(JSON.stringify(config) !== JSON.stringify(originalConfig))
-    } else {
-      setHasChanges(sourceCode !== originalSourceCode)
-    }
-  }, [config, originalConfig, sourceCode, originalSourceCode, editMode])
-
-  // 处理字段变化
-  const handleFieldChange = (sectionName: string, fieldName: string, value: unknown) => {
-    setConfig(prev => setNestedField(prev, sectionName, fieldName, value))
-  }
-
-  // 保存配置
-  const handleSave = async (): Promise<boolean> => {
-    setSaving(true)
-    try {
-      if (editMode === 'source') {
-        // 源代码模式：先验证 TOML 格式
-        try {
-          parseToml(sourceCode)
-        } catch (error) {
-          setHasTomlError(true)
-          toast({
-            title: 'TOML 格式错误',
-            description: error instanceof Error ? error.message : '无法解析 TOML 配置，请检查语法',
-            variant: 'destructive'
-          })
-          setSaving(false)
-          return false
-        }
-        
-        // 格式正确，保存原始配置
-        await updatePluginConfigRaw(plugin.id, sourceCode)
-        setOriginalSourceCode(sourceCode)
-        setHasTomlError(false)
-      } else {
-        // 可视化模式
-        await updatePluginConfig(plugin.id, config)
-        setOriginalConfig(JSON.parse(JSON.stringify(config)))
-      }
-      
-      toast({
-        title: '配置已保存',
-        description: '更改将在插件重新加载后生效'
-      })
-      return true
-    } catch (error) {
-      toast({
-        title: '保存失败',
-        description: error instanceof Error ? error.message : '未知错误',
-        variant: 'destructive'
-      })
-      return false
-    } finally {
-      setSaving(false)
-    }
-  }
-
-  const handleBack = () => {
-    if (!hasChanges) {
-      onBack()
-      return
-    }
-    setInternalLeavePromptOpen(true)
-  }
-
-  const closeLeavePrompt = () => {
-    if (navigationBlocker.status === 'blocked') {
-      navigationBlocker.reset?.()
-    }
-    setInternalLeavePromptOpen(false)
-  }
-
-  const leaveWithoutSaving = () => {
-    if (internalLeavePromptOpen) {
-      setInternalLeavePromptOpen(false)
-      onBack()
-      return
-    }
-    navigationBlocker.proceed?.()
-  }
-
-  const saveAndLeave = async () => {
-    const saved = await handleSave()
-    if (!saved) {
-      return
-    }
-    if (internalLeavePromptOpen) {
-      setInternalLeavePromptOpen(false)
-      onBack()
-      return
-    }
-    navigationBlocker.proceed?.()
-  }
-
-  // 重置配置
-  const handleReset = async () => {
-    try {
-      const resetResult = await resetPluginConfig(plugin.id)
-      if (!resetResult.success) {
-        toast({
-          title: '重置失败',
-          description: resetResult.error,
-          variant: 'destructive'
-        })
-        return
-      }
-      toast({
-        title: '配置已重置',
-        description: '下次加载插件时将使用默认配置'
-      })
-      setResetDialogOpen(false)
-      loadConfig()
-    } catch (error) {
-      toast({
-        title: '重置失败',
-        description: error instanceof Error ? error.message : '未知错误',
-        variant: 'destructive'
-      })
-    }
-  }
-
-  // 切换启用状态
-  const handleToggle = async () => {
-    try {
-      const toggleResult = await togglePlugin(plugin.id)
-      if (!toggleResult.success) {
-        toast({
-          title: '切换失败',
-          description: toggleResult.error,
-          variant: 'destructive'
-        })
-        return
-      }
-      toast({
-        title: toggleResult.data.message,
-        description: toggleResult.data.note
-      })
-      loadConfig()
-    } catch (error) {
-      toast({
-        title: '切换状态失败',
-        description: error instanceof Error ? error.message : '未知错误',
-        variant: 'destructive'
-      })
-    }
-  }
+  const {
+    editMode,
+    setEditMode,
+    pluginPageTab,
+    setPluginPageTab,
+    activeConfigTab,
+    handleConfigTabChange,
+    schema,
+    config,
+    sourceCode,
+    handleSourceCodeChange,
+    handleFieldChange,
+    loading,
+    saving,
+    hasChanges,
+    hasTomlError,
+    handleSave,
+    handleReset,
+    handleToggle,
+    resetDialogOpen,
+    setResetDialogOpen,
+    navigationBlocker,
+    internalLeavePromptOpen,
+    handleBack,
+    closeLeavePrompt,
+    leaveWithoutSaving,
+    saveAndLeave,
+  } = usePluginConfigEditor({ plugin, onBack, initialTab })
 
   if (loading) {
     return (
@@ -740,12 +467,6 @@ function PluginConfigEditor({ plugin, onBack, initialTab }: PluginConfigEditorPr
   const selectedConfigTab = schemaTabs.some((tab) => tab.id === activeConfigTab)
     ? activeConfigTab
     : schemaTabs[0]?.id
-
-  const handleConfigTabChange = (nextTab: string) => {
-    setActiveConfigTab(nextTab)
-    const params = new URLSearchParams({ plugin: plugin.id, tab: nextTab })
-    window.history.replaceState(null, '', `/plugin-config?${params.toString()}`)
-  }
 
   // 获取当前启用状态
   const isEnabled = (config.plugin as Record<string, unknown>)?.enabled !== false
@@ -889,12 +610,7 @@ function PluginConfigEditor({ plugin, onBack, initialTab }: PluginConfigEditorPr
 
               <CodeEditor
                 value={sourceCode}
-                onChange={(value) => {
-                  setSourceCode(value)
-                  if (hasTomlError) {
-                    setHasTomlError(false)
-                  }
-                }}
+                onChange={handleSourceCodeChange}
                 language="toml"
                 height="calc(100vh - 350px)"
                 minHeight="500px"
@@ -1055,33 +771,6 @@ function PluginConfigEditor({ plugin, onBack, initialTab }: PluginConfigEditorPr
 }
 
 // 主页面组件 - 包装 RestartProvider
-function getInitialPluginConfigTarget(): { pluginId: string | null; tabId: string | null } {
-  if (typeof window === 'undefined') {
-    return { pluginId: null, tabId: null }
-  }
-
-  const params = new URLSearchParams(window.location.search)
-  return {
-    pluginId: params.get('plugin'),
-    tabId: params.get('tab'),
-  }
-}
-
-function comparePluginVersions(currentVersion: string, latestVersion: string): number {
-  const currentParts = currentVersion.trim().split('.').map(part => Number.parseInt(part, 10) || 0)
-  const latestParts = latestVersion.trim().split('.').map(part => Number.parseInt(part, 10) || 0)
-  const maxLength = Math.max(currentParts.length, latestParts.length)
-
-  for (let index = 0; index < maxLength; index++) {
-    const currentPart = currentParts[index] || 0
-    const latestPart = latestParts[index] || 0
-    if (latestPart > currentPart) return 1
-    if (latestPart < currentPart) return -1
-  }
-
-  return 0
-}
-
 export function PluginConfigPage() {
   return (
     <RestartProvider>
@@ -1092,525 +781,70 @@ export function PluginConfigPage() {
 
 // 内部组件：实际内容
 function PluginConfigPageContent() {
-  const { toast } = useToast()
   const { themeConfig } = useTheme()
-  const initialTarget = getInitialPluginConfigTarget()
-  const [plugins, setPlugins] = useState<InstalledPlugin[]>([])
-  const [loading, setLoading] = useState(true)
-  const [searchQuery, setSearchQuery] = useState('')
-  const [showUpdateOnly, setShowUpdateOnly] = useState(false)
-  const [selectedPlugin, setSelectedPlugin] = useState<InstalledPlugin | null>(null)
-  const [selectedPluginTab, setSelectedPluginTab] = useState<string | undefined>(initialTarget.tabId ?? undefined)
-  const [actingPluginId, setActingPluginId] = useState<string | null>(null)
-  const [deleteDialogOpen, setDeleteDialogOpen] = useState(false)
-  const [deletingPlugin, setDeletingPlugin] = useState<InstalledPlugin | null>(null)
-  const [deleteProgress, setDeleteProgress] = useState<PluginLoadProgress | null>(null)
-  const [updateDialogOpen, setUpdateDialogOpen] = useState(false)
-  const [updatingPlugin, setUpdatingPlugin] = useState<InstalledPlugin | null>(null)
-  const [updateProgress, setUpdateProgress] = useState<PluginLoadProgress | null>(null)
-  const [marketPluginsById, setMarketPluginsById] = useState<Record<string, PluginInfo>>({})
-  const [checkingUpdates, setCheckingUpdates] = useState(false)
 
-  const openPluginConfig = (plugin: InstalledPlugin, tabId?: string | null) => {
-    setSelectedPlugin(plugin)
-    setSelectedPluginTab(tabId ?? undefined)
-    const params = new URLSearchParams({ plugin: plugin.id })
-    if (tabId) {
-      params.set('tab', tabId)
-    }
-    window.history.replaceState(null, '', `/plugin-config?${params.toString()}`)
-  }
+  const {
+    plugins,
+    loading,
+    selectedPlugin,
+    selectedPluginTab,
+    openPluginConfig,
+    closePluginConfig,
+    loadPlugins,
+    searchQuery,
+    setSearchQuery,
+    showUpdateOnly,
+    setShowUpdateOnly,
+    visiblePlugins,
+    actingPluginId,
+    setActingPluginId,
+    performTogglePlugin,
+    checkingUpdates,
+    getPluginUpdateState,
+    getPluginRepositoryUrl,
+    isPluginDisabled,
+    getPluginStatusBarClassName,
+    getPluginStatusLabel,
+    getPluginStatusMeta,
+    installedCount,
+    disabledCount,
+    loadingCount,
+    circuitOpenCount,
+    loadFailedCount,
+    enabledCount,
+    loadSuccessCount,
+    loadSuccessPercent,
+    loadFailedPercent,
+    loadingPercent,
+    circuitPercent,
+    showsCircuitSummary,
+    modernLoadSummaryLabel,
+    futureRetroPluginSummaryLabel,
+  } = usePluginList()
 
-  const closePluginConfig = () => {
-    setSelectedPlugin(null)
-    setSelectedPluginTab(undefined)
-    window.history.replaceState(null, '', '/plugin-config')
-  }
-
-  // 加载插件列表
-  const loadPlugins = async () => {
-    setLoading(true)
-    try {
-      const installedResult = await getInstalledPlugins()
-      if (!installedResult.success) {
-        toast({
-          title: '加载插件列表失败',
-          description: installedResult.error,
-          variant: 'destructive'
-        })
-        return
-      }
-      setPlugins(installedResult.data)
-      if (!selectedPlugin && initialTarget.pluginId) {
-        const targetPlugin = installedResult.data.find((plugin) => plugin.id === initialTarget.pluginId)
-        if (targetPlugin) {
-          openPluginConfig(targetPlugin, initialTarget.tabId)
-        }
-      }
-    } catch (error) {
-      toast({
-        title: '加载插件列表失败',
-        description: error instanceof Error ? error.message : '未知错误',
-        variant: 'destructive'
-      })
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  const checkPluginUpdates = async () => {
-    setCheckingUpdates(true)
-    try {
-      const marketResult = await fetchPluginList()
-      if (!marketResult.success) {
-        console.warn('加载插件市场版本信息失败:', marketResult.error)
-        setMarketPluginsById({})
-        return
-      }
-
-      const nextMarketPluginsById: Record<string, PluginInfo> = {}
-      for (const marketPlugin of marketResult.data) {
-        nextMarketPluginsById[marketPlugin.id] = marketPlugin
-        if (marketPlugin.manifest.id) {
-          nextMarketPluginsById[marketPlugin.manifest.id] = marketPlugin
-        }
-      }
-      setMarketPluginsById(nextMarketPluginsById)
-    } catch (error) {
-      console.warn('加载插件市场版本信息失败:', error)
-      setMarketPluginsById({})
-    } finally {
-      setCheckingUpdates(false)
-    }
-  }
-
-  useEffect(() => {
-    loadPlugins()
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  useEffect(() => {
-    if (!loading) {
-      void checkPluginUpdates()
-    }
-  }, [loading])
-
-  useEffect(() => {
-    let unsubscribe: (() => Promise<void>) | null = null
-    let disposed = false
-
-    void pluginProgressClient.subscribe((progress) => {
-      if (disposed) {
-        return
-      }
-      if (progress.operation === 'uninstall' && deletingPlugin && progress.plugin_id === deletingPlugin.id) {
-        setDeleteProgress(progress)
-      }
-      if (progress.operation === 'update' && updatingPlugin && progress.plugin_id === updatingPlugin.id) {
-        setUpdateProgress(progress)
-      }
-    }).then((cleanup) => {
-      if (disposed) {
-        void cleanup()
-        return
-      }
-      unsubscribe = cleanup
-    })
-
-    return () => {
-      disposed = true
-      if (unsubscribe) {
-        void unsubscribe()
-      }
-    }
-  }, [deletingPlugin, updatingPlugin])
-
-  // 过滤插件
-  const filteredPlugins = plugins.filter(plugin => {
-    const query = searchQuery.toLowerCase()
-    return (
-      plugin.id.toLowerCase().includes(query) ||
-      plugin.manifest.name.toLowerCase().includes(query) ||
-      plugin.manifest.description?.toLowerCase().includes(query)
-    )
+  const {
+    deleteDialogOpen,
+    setDeleteDialogOpen,
+    deletingPlugin,
+    deleteProgress,
+    openDeletePluginDialog,
+    closeDeletePluginDialog,
+    handleConfirmDeletePlugin,
+    updateDialogOpen,
+    setUpdateDialogOpen,
+    updatingPlugin,
+    updateProgress,
+    openUpdatePluginDialog,
+    closeUpdatePluginDialog,
+    handleConfirmUpdatePlugin,
+  } = usePluginLifecycle({
+    getPluginRepositoryUrl,
+    onChanged: loadPlugins,
+    setActingPluginId,
   })
-  
-  // 去重：如果有重复的 plugin.id，只保留第一个
-  const uniqueFilteredPlugins = filteredPlugins.filter((plugin, index, self) =>
-    index === self.findIndex((p) => p.id === plugin.id)
-  )
 
-  // 统计数据
-  const isPluginDisabled = (plugin: InstalledPlugin) => plugin.disabled === true || plugin.enabled === false
-  const isPluginLoadSuccess = (plugin: InstalledPlugin) => !isPluginDisabled(plugin) && (
-    plugin.load_status === 'success' || plugin.loaded === true
-  )
-  const isPluginLoading = (plugin: InstalledPlugin) => !isPluginDisabled(plugin) && plugin.load_status === 'loading'
-  const isPluginCircuitOpen = (plugin: InstalledPlugin) => !isPluginDisabled(plugin) && plugin.circuit_status?.state === 'open'
-  const isPluginCircuitHalfOpen = (plugin: InstalledPlugin) => !isPluginDisabled(plugin) && plugin.circuit_status?.state === 'half_open'
-  const isPluginCircuitActive = (plugin: InstalledPlugin) => isPluginCircuitOpen(plugin) || isPluginCircuitHalfOpen(plugin)
-  const isPluginLoadFailed = (plugin: InstalledPlugin) => (
-    !isPluginDisabled(plugin)
-    && !isPluginLoading(plugin)
-    && !isPluginLoadSuccess(plugin)
-  )
-  const installedCount = plugins.length
-  const disabledCount = plugins.filter(isPluginDisabled).length
-  const loadSuccessCount = plugins.filter(isPluginLoadSuccess).length
-  const loadingCount = plugins.filter(isPluginLoading).length
-  const circuitOpenCount = plugins.filter(isPluginCircuitOpen).length
-  const circuitActiveCount = plugins.filter(isPluginCircuitActive).length
-  const loadFailedCount = plugins.filter(isPluginLoadFailed).length
-  const enabledCount = installedCount - disabledCount
-  const loadTotalCount = loadSuccessCount + loadFailedCount + loadingCount + circuitActiveCount
-  const loadSuccessPercent = loadTotalCount > 0 ? (loadSuccessCount / loadTotalCount) * 100 : 0
-  const loadFailedPercent = loadTotalCount > 0 ? (loadFailedCount / loadTotalCount) * 100 : 0
-  const loadingPercent = loadTotalCount > 0 ? (loadingCount / loadTotalCount) * 100 : 0
-  const circuitPercent = loadTotalCount > 0 ? (circuitActiveCount / loadTotalCount) * 100 : 0
-  const showsCircuitSummary = circuitOpenCount > 0
-  const modernLoadSummaryLabel = [
-    `加载成功 ${loadSuccessCount} 个`,
-    `加载中 ${loadingCount} 个`,
-    showsCircuitSummary ? `熔断中 ${circuitOpenCount} 个` : '',
-    `加载失败 ${loadFailedCount} 个`,
-  ].filter(Boolean).join('，')
-  const futureRetroPluginSummaryLabel = [
-    `已安装 ${installedCount} 个插件`,
-    `已启用 ${enabledCount} 个`,
-    `已禁用 ${disabledCount} 个`,
-    `加载中 ${loadingCount} 个`,
-    showsCircuitSummary ? `熔断中 ${circuitOpenCount} 个` : '',
-    `启动失败 ${loadFailedCount} 个`,
-  ].filter(Boolean).join('，')
   const isModernDashboardStyle = themeConfig.dashboardStyle === 'modern'
   const isFutureRetroDashboardStyle = themeConfig.dashboardStyle === 'future-retro'
-  const getPluginStatusBarClassName = (plugin: InstalledPlugin) => {
-    if (isPluginDisabled(plugin)) {
-      return 'bg-muted-foreground/45'
-    }
-    if (isPluginCircuitOpen(plugin)) {
-      return 'bg-orange-500'
-    }
-    if (isPluginCircuitHalfOpen(plugin)) {
-      return 'bg-yellow-500'
-    }
-    if (isPluginLoading(plugin)) {
-      return 'bg-sky-500'
-    }
-    if (isPluginLoadFailed(plugin)) {
-      return 'bg-red-500'
-    }
-    return 'bg-emerald-500'
-  }
-  const getPluginStatusLabel = (plugin: InstalledPlugin) => {
-    if (isPluginDisabled(plugin)) {
-      return '已禁用'
-    }
-    if (isPluginCircuitOpen(plugin)) {
-      const remainingSec = Math.ceil(plugin.circuit_status?.remaining_sec ?? 0)
-      return remainingSec > 0 ? `熔断中 ${remainingSec}s` : '熔断中'
-    }
-    if (isPluginCircuitHalfOpen(plugin)) {
-      return '半开测试'
-    }
-    if (isPluginLoading(plugin)) {
-      return '加载中'
-    }
-    if (isPluginLoadFailed(plugin)) {
-      return '启动失败'
-    }
-    return '已启用'
-  }
-  const getPluginStatusMeta = (plugin: InstalledPlugin): PluginStatusMeta => {
-    if (isPluginDisabled(plugin)) {
-      return { dotClassName: 'bg-muted-foreground/45', label: '已禁用' }
-    }
-    if (isPluginCircuitOpen(plugin)) {
-      const remainingSec = Math.ceil(plugin.circuit_status?.remaining_sec ?? 0)
-      return {
-        dotClassName: 'bg-orange-500',
-        label: remainingSec > 0 ? `熔断中 ${remainingSec}s` : '熔断中',
-        badgeClassName: 'border-orange-600 text-orange-600',
-        icon: 'circuit' as const,
-      }
-    }
-    if (isPluginCircuitHalfOpen(plugin)) {
-      return {
-        dotClassName: 'bg-yellow-500',
-        label: '半开测试',
-        badgeClassName: 'border-yellow-600 text-yellow-700',
-        icon: 'warning' as const,
-      }
-    }
-    if (isPluginLoading(plugin)) {
-      return {
-        dotClassName: 'bg-sky-500',
-        label: '加载中',
-        badgeClassName: 'border-sky-600 text-sky-600',
-        icon: 'loading' as const,
-      }
-    }
-    if (isPluginLoadSuccess(plugin)) {
-      return { dotClassName: 'bg-emerald-500', label: '加载成功', showsBadge: false }
-    }
-    return {
-      dotClassName: 'bg-red-500',
-      label: '加载失败',
-      badgeClassName: 'border-red-600 text-red-600',
-      icon: 'warning' as const,
-    }
-  }
-  const getPluginRepositoryUrl = (plugin: InstalledPlugin): string | undefined => {
-    const marketPlugin = marketPluginsById[plugin.id] || (plugin.manifest.id ? marketPluginsById[plugin.manifest.id] : undefined)
-    const urls = plugin.manifest.urls as { repository?: string } | undefined
-    return plugin.manifest.repository_url || urls?.repository || marketPlugin?.manifest.repository_url || marketPlugin?.manifest.urls?.repository
-  }
-  const getPluginUpdateState = (plugin: InstalledPlugin): { canUpdate: boolean; hasUpdate: boolean; title?: string } => {
-    if (checkingUpdates) {
-      return { canUpdate: false, hasUpdate: false, title: '正在检查更新' }
-    }
-
-    const marketPlugin = marketPluginsById[plugin.id] || (plugin.manifest.id ? marketPluginsById[plugin.manifest.id] : undefined)
-    if (!marketPlugin) {
-      return { canUpdate: false, hasUpdate: false, title: '插件市场中没有找到该插件，无法判断新版本' }
-    }
-
-    if (!getPluginRepositoryUrl(plugin)) {
-      return { canUpdate: false, hasUpdate: false, title: '插件清单中没有仓库地址，无法更新/升级' }
-    }
-
-    const currentVersion = plugin.manifest.version
-    const latestVersion = marketPlugin.manifest.version
-    if (comparePluginVersions(currentVersion, latestVersion) <= 0) {
-      return { canUpdate: false, hasUpdate: false, title: '当前已是最新版本' }
-    }
-
-    return { canUpdate: true, hasUpdate: true, title: `发现新版本 v${latestVersion}` }
-  }
-  const visiblePlugins = showUpdateOnly
-    ? uniqueFilteredPlugins.filter((plugin) => getPluginUpdateState(plugin).hasUpdate)
-    : uniqueFilteredPlugins
-
-  const stopPluginActionEvent = (event: React.MouseEvent<HTMLElement>) => {
-    event.preventDefault()
-    event.stopPropagation()
-  }
-
-  const performTogglePlugin = async (plugin: InstalledPlugin) => {
-    setActingPluginId(plugin.id)
-    try {
-      const toggleResult = await togglePlugin(plugin.id)
-      if (!toggleResult.success) {
-        toast({
-          title: '切换插件状态失败',
-          description: toggleResult.error,
-          variant: 'destructive'
-        })
-        return
-      }
-
-      toast({
-        title: toggleResult.data.enabled ? '插件已启动' : '插件已关闭',
-        description: toggleResult.data.message || `${plugin.manifest.name} 状态已更新`
-      })
-      await loadPlugins()
-    } catch (error) {
-      toast({
-        title: '切换插件状态失败',
-        description: error instanceof Error ? error.message : '未知错误',
-        variant: 'destructive'
-      })
-    } finally {
-      setActingPluginId(null)
-    }
-  }
-
-  const openUpdatePluginDialog = (plugin: InstalledPlugin, event: React.MouseEvent<HTMLButtonElement>) => {
-    stopPluginActionEvent(event)
-    setUpdatingPlugin(plugin)
-    setUpdateProgress(null)
-    setUpdateDialogOpen(true)
-  }
-
-  const closeUpdatePluginDialog = () => {
-    if (updateProgress?.stage === 'loading') {
-      return
-    }
-    setUpdateDialogOpen(false)
-    setUpdatingPlugin(null)
-    setUpdateProgress(null)
-  }
-
-  const handleConfirmUpdatePlugin = async () => {
-    if (!updatingPlugin) return
-
-    const repositoryUrl = getPluginRepositoryUrl(updatingPlugin)
-    if (!repositoryUrl) {
-      setUpdateProgress({
-        operation: 'update',
-        stage: 'error',
-        progress: 0,
-        message: '插件清单中没有仓库地址，无法更新/升级',
-        error: '插件清单中没有仓库地址，无法更新/升级',
-        plugin_id: updatingPlugin.id,
-        total_plugins: 1,
-        loaded_plugins: 0,
-      })
-      return
-    }
-
-    setActingPluginId(updatingPlugin.id)
-    setUpdateProgress({
-      operation: 'update',
-      stage: 'loading',
-      progress: 0,
-      message: `正在准备更新 ${updatingPlugin.manifest.name}`,
-      plugin_id: updatingPlugin.id,
-      total_plugins: 1,
-      loaded_plugins: 0,
-    })
-    try {
-      const updateResult = await updatePlugin(updatingPlugin.id, repositoryUrl, 'main')
-      if (!updateResult.success) {
-        setUpdateProgress({
-          operation: 'update',
-          stage: 'error',
-          progress: 0,
-          message: updateResult.error || '更新插件失败',
-          error: updateResult.error || '更新插件失败',
-          plugin_id: updatingPlugin.id,
-          total_plugins: 1,
-          loaded_plugins: 0,
-        })
-        toast({
-          title: '更新插件失败',
-          description: updateResult.error,
-          variant: 'destructive'
-        })
-        return
-      }
-
-      toast({
-        title: '更新插件成功',
-        description: `${updatingPlugin.manifest.name} 已完成更新/升级`
-      })
-      setUpdateProgress({
-        operation: 'update',
-        stage: 'success',
-        progress: 100,
-        message: `${updatingPlugin.manifest.name} 已完成更新/升级`,
-        plugin_id: updatingPlugin.id,
-        total_plugins: 1,
-        loaded_plugins: 1,
-      })
-      await loadPlugins()
-    } catch (error) {
-      const errorMessage = error instanceof Error ? error.message : '未知错误'
-      setUpdateProgress({
-        operation: 'update',
-        stage: 'error',
-        progress: 0,
-        message: errorMessage,
-        error: errorMessage,
-        plugin_id: updatingPlugin.id,
-        total_plugins: 1,
-        loaded_plugins: 0,
-      })
-      toast({
-        title: '更新插件失败',
-        description: errorMessage,
-        variant: 'destructive'
-      })
-    } finally {
-      setActingPluginId(null)
-    }
-  }
-
-  const openDeletePluginDialog = (plugin: InstalledPlugin, event: React.MouseEvent<HTMLButtonElement>) => {
-    stopPluginActionEvent(event)
-    setDeletingPlugin(plugin)
-    setDeleteProgress(null)
-    setDeleteDialogOpen(true)
-  }
-
-  const closeDeletePluginDialog = () => {
-    if (deleteProgress?.stage === 'loading') {
-      return
-    }
-    setDeleteDialogOpen(false)
-    setDeletingPlugin(null)
-    setDeleteProgress(null)
-  }
-
-  const handleConfirmDeletePlugin = async () => {
-    if (!deletingPlugin) return
-
-    setActingPluginId(deletingPlugin.id)
-    setDeleteProgress({
-      operation: 'uninstall',
-      stage: 'loading',
-      progress: 0,
-      message: `正在准备删除 ${deletingPlugin.manifest.name}`,
-      plugin_id: deletingPlugin.id,
-      total_plugins: 1,
-      loaded_plugins: 0,
-    })
-    try {
-      const uninstallResult = await uninstallPlugin(deletingPlugin.id)
-      if (!uninstallResult.success) {
-        setDeleteProgress({
-          operation: 'uninstall',
-          stage: 'error',
-          progress: 0,
-          message: uninstallResult.error || '删除插件失败',
-          error: uninstallResult.error || '删除插件失败',
-          plugin_id: deletingPlugin.id,
-          total_plugins: 1,
-          loaded_plugins: 0,
-        })
-        toast({
-          title: '删除插件失败',
-          description: uninstallResult.error,
-          variant: 'destructive'
-        })
-        return
-      }
-
-      toast({
-        title: '删除插件成功',
-        description: `${deletingPlugin.manifest.name} 已删除`
-      })
-      setDeleteProgress({
-        operation: 'uninstall',
-        stage: 'success',
-        progress: 100,
-        message: `${deletingPlugin.manifest.name} 已删除`,
-        plugin_id: deletingPlugin.id,
-        total_plugins: 1,
-        loaded_plugins: 1,
-      })
-      await loadPlugins()
-    } catch (error) {
-      const errorMessage = error instanceof Error ? error.message : '未知错误'
-      setDeleteProgress({
-        operation: 'uninstall',
-        stage: 'error',
-        progress: 0,
-        message: errorMessage,
-        error: errorMessage,
-        plugin_id: deletingPlugin.id,
-        total_plugins: 1,
-        loaded_plugins: 0,
-      })
-      toast({
-        title: '删除插件失败',
-        description: errorMessage,
-        variant: 'destructive'
-      })
-    } finally {
-      setActingPluginId(null)
-    }
-  }
 
   // 如果选中了插件，显示配置编辑器
   if (selectedPlugin) {
